requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
python-dotenv>=0.21.0
typing-extensions>=4.12.2
annotated-types>=0.6.0
pymysql==1.1.0
alembic==1.12.0
pytest==7.4.0
httpx==0.24.1
redis==4.6.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pillow==10.0.0
requests==2.31.0
alibabacloud_dysmsapi20170525==3.1.1